requirements.txt
oracledb==2.1.0
python-telegram-bot==21.0.1
python-telegram-bot[job-queue]
peewee